refactor: Aligning line spacing
