+++ src/App.tsx
@@ -1,21 +1,8 @@
-import Sidebar from "@/components/Home/Sidebar";
-
-function App() {
-  return (
-    <div className='flex'>
-      <Sidebar />
-    </div>
-  )
-}
-
-export default App;
-
-/*
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
-import LoginForm from './components/LoginForm';
-import Home from './components/Home';
-import { useStore } from './store';
+import Home from '@/components/Home';
+import Login from '@/components/Login';
+import { useStore } from '@/stores/login';
 
 const App: React.FC = () => {
   const { isLoggedIn } = useStore();
@@ -24,15 +11,13 @@ const App: React.FC = () => {
     <Router>
       <div className="min-h-screen bg-gray-900">
         <Routes>
-          <Route path="/login" element={
-            isLoggedIn ? <Navigate to="/home" /> : <LoginForm />
-          } />
-          <Route path="/home" element={
-            isLoggedIn ? <Home /> : <Navigate to="/login" />
-          } />
+          <Route path="/login" element={isLoggedIn ? <Navigate to="/home" /> : <Login />} />
+          <Route path="/home" element={isLoggedIn ? <Home /> : <Navigate to="/login" />} />
           <Route path="/" element={<Navigate to="/login" />} />
         </Routes>
       </div>
     </Router>
   );
-}*/
+};
+
+export default App;
