+++ src/components/Home/Sidebar/icons/HomeIcon.tsx
@@ -4,7 +4,14 @@ import React from 'react';
  * @todo 임시 아이콘입니다. 프로젝트 아이콘이 만들어지면 변경
  */
 export const HomeIcon: React.FC = () => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" className="bi bi-house" viewBox="0 0 16 16">
-        <path d="M8.354 0.146a.5.5 0 0 1 .292.086l7 4a.5.5 0 0 1 0 .868l-1 1A.5.5 0 0 1 14 6.5V14a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V6.5a.5.5 0 0 1-.146-.354l-1-1a.5.5 0 0 1 0-.868l7-4a.5.5 0 0 1 .292-.086zM3 14h10V6.5l-5-2.857L3 6.5V14z" />
-    </svg>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="20"
+    height="20"
+    fill="currentColor"
+    className="bi bi-house"
+    viewBox="0 0 16 16"
+  >
+    <path d="M8.354 0.146a.5.5 0 0 1 .292.086l7 4a.5.5 0 0 1 0 .868l-1 1A.5.5 0 0 1 14 6.5V14a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V6.5a.5.5 0 0 1-.146-.354l-1-1a.5.5 0 0 1 0-.868l7-4a.5.5 0 0 1 .292-.086zM3 14h10V6.5l-5-2.857L3 6.5V14z" />
+  </svg>
 );
