+++ src/components/Home/index.tsx
@@ -1,0 +1,11 @@
+import Sidebar from '@/components/Home/Sidebar';
+
+const Home: React.FC = () => {
+  return (
+    <div className="flex">
+      <Sidebar />
+    </div>
+  );
+};
+
+export default Home;
