+++ src/index.css
@@ -19,7 +19,7 @@
     hover:rounded-xl rounded-3xl
     transition-all duration-100 ease-linear
     cursor-pointer shadow-lg;
-  }  
+  }
 
   .sidebar-tooltip {
     @apply absolute w-auto p-2 m-2 min-w-max left-14 rounded-md shadow-md
